Shape

+++ team1/src/shape/Circle.java
@@ -2,9 +2,9 @@ package shape;
 
 public class Circle implements Shape {
 
-	private int radius;
+	private double radius;
 	
-	public Circle(int radius) {
+	public Circle(double radius) {
 		this.radius = radius;
 	}
